Implemented a simple version of Conoways game of life for Dawn

+++ Kernel/headers/input/mouse.h
@@ -2,8 +2,8 @@
 #define _MOUSE_DEF_H_
 
 typedef struct {
-	uint8 i_byte;
-	int8 mouse_x, mouse_y;
+	unsigned int i_byte;
+	int mouse_x, mouse_y;
 } mouse_input_t; //Structure to contain data sent from a ps/2 mouse
 
 #endif //_MOUSE_DEF_H_
